+++ src/main/java/me/khajiitos/jackseconomy/data/StockMarketManager.java
@@ -1,12 +1,11 @@
 package me.khajiitos.jackseconomy.data;
 
-import com.google.gson.JsonObject;
 import me.khajiitos.jackseconomy.JacksEconomy;
+import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.level.storage.LevelResource;
 
 public class StockMarketManager {
 	private static DataHandler dataHandler;
-	private static JsonObject data;
 	private static long lastUpdated;
 
 	public static void load() {
@@ -15,39 +14,44 @@ public class StockMarketManager {
 						.resolve("data/jackseconomy/stock-market.dat")
 						.toFile()
 		);
+		resetData();
 		if (dataHandler.DATA_FILE.exists()) {
-			data = dataHandler.loadAsJson();
-			lastUpdated = data.has("lastUpdated") ? data.get("lastUpdated").getAsLong() : 0;
-		} else {
-			data = new JsonObject();
-			data.addProperty("lastUpdated", 0);
-
-			lastUpdated = 0;
-
-			dataHandler.save(data);
-		}
-		JacksEconomy.server.addTickable(StockMarketManager::tick);
+			CompoundTag data = dataHandler.loadAsNbt();
+			lastUpdated = data.contains("lastUpdated") ? data.getLong("lastUpdated") : 0;
+		} else save();
 	}
+
 	public static void save() {
-		data.addProperty("lastUpdated", lastUpdated);
+		CompoundTag data = new CompoundTag();
+		data.putLong("lastUpdated", lastUpdated);
 
 		dataHandler.save(data);
 	}
+
+	public static void resetData() {
+		lastUpdated = 0;
+	}
+
 	public static void tick() {
 		if (getDay() > lastUpdated) {
 			lastUpdated = getDay();
-			update();
+			tryUpdate();
 		}
 	}
+
 	private static long getDay() {
 		return JacksEconomy.server.overworld().getDayTime() / 24000;
 	}
-	private static void update() {
+
+	private static void tryUpdate() {
 		JacksEconomy.LOGGER.info("Updating stock market...");
 		try {
+			update();
 			JacksEconomy.LOGGER.info("Updated stock market!");
 		} catch (Exception ignored) {
 			JacksEconomy.LOGGER.error("Error updating stock market!");
 		}
 	}
+
+	private static void update() {}
 }